feat: Nacharbeiten zu Automatisches-Update-Fix abgeschlossen (Code-Review-Befunde behoben)
Behebt die verbleibenden Code-Review-Befunde aus dem vorigen Durchlauf:
gemeinsame TestWebHostEnvironment- und UpdateStatusTestData-Fixtures statt
Duplikaten, Logging im Lock-Zeitstempel-Fallback von UpdateFileStore,
nicht-blockierender Testcode sowie Aufteilung der UpdateMetadataAndPlatformTests
in fokussierte Testklassen. Feature-Dokumentation entfernt (bereits über
Planungscommit in der Historie erhalten).
