--- Docs/features/task/issue-206-a764a3ecd8074cc7a2c0fef9bc6e2a1d-automatisches-update-repariere/todo.md
@@ -1,25 +1,0 @@
-# Aufgabenliste – Anforderungsbearbeitung
-
-Branch: `task/issue-206-a764a3ecd8074cc7a2c0fef9bc6e2a1d-automatisches-update-repariere`
-
-| Status | Schritt | Beschreibung | Artefakt |
-|--------|---------|--------------|----------|
-| [x] | 1 | Branch-Name ermitteln | – |
-| [x] | 2 | Verzeichnisstruktur vorbereiten | `docs/features/{branchname}/` |
-| [x] | – | Einstiegspunkt ermitteln | – |
-| [x] | 3 | Anforderung übersetzen (Unteragent) | `requirement.md` |
-| [x] | 4 | Bestandsaufnahme (Unteragent) | `inventory.md`, `inventory/` |
-| [x] | 5 | Umsetzungsplanung (Unteragent) | `plan.md` |
-| [x] | 5a | Offene Punkte prüfen und ggf. Planung wiederholen | `plan.md` (aktualisiert) |
-| [x] | 5b | Planungscommit | – |
-| [x] | 6 | Implementierung (Unteragent) | Codeänderungen |
-| [x] | 7 | Plan-Review (Unteragent, bedingt) | `review.md` |
-| [x] | 8 | Code-Review (Unteragent) | `review-code.md` |
-| [x] | 8b | Tests ausführen (Unteragent) | `test-results.md` |
-| [x] | – | Iteration oder Abschluss entscheiden | – |
-| [x] | 8a | Folgeaufgaben dokumentieren (bei Schleifenabbruch) | `continue.md` |
-| [x] | 9 | Dokumentation erstellen (Unteragent) | `docs/help/` |
-| [x] | 9b | README aktualisieren (Unteragent) | `README.md` |
-| [ ] | 10 | Nacharbeiten abschließen (offene Punkte aus `continue.md`) | `continue-done.md` |
-| [ ] | – | Feature-Verzeichnis löschen | – |
-| [ ] | – | Commit durchführen | – |
